Database changed and User auth upgraded
Database -> sqlite3
User's password encrypted.
Can register new user.
