+++ setupbox/web_server/templates/base.html
@@ -36,14 +36,11 @@
         $.contextMenu({
             selector: '.files',
             callback: function(key, options) {
-                //var m = "clicked: " + key;
-                // window.console && console.log(m) || alert(m);
-
                 $.post('/'+key,{ path:$(this.context).attr('id'),
                     directory_root:$("#folder_path").val() });
             },
             items: {
-                "rename": {name: "이름 바꾸기", icon: "rename"},
+                "rename": {name: "이름 바꾸기", icon: "edit"},
                 "delete": {name: "삭제하기", icon: "delete"}
             }
         });
@@ -53,14 +50,11 @@
         $.contextMenu({
             selector: '.folders',
             callback: function(key, options) {
-                //var m = "clicked: " + key;
-                // window.console && console.log(m) || alert(m);
-
                 $.post('/'+key,{ path:$(this.context).attr('id'),
                     directory_root:$("#folder_path").val() });
             },
             items: {
-                "rename": {name: "이름 바꾸기", icon: "rename"},
+                "rename": {name: "이름 바꾸기", icon: "edit"},
                 "delete": {name: "삭제하기", icon: "delete"}
             }
         });
@@ -87,6 +81,8 @@
                     <ul class="nav navbar-nav">
                         {# <li><a href="{{ url_for('security.register') }}">회원가입</a></li> #}
                         <li><a href="{{ url_for('security.login') }}">로그인</a></li>
+                        <li><a href="{{ url_for('security.forgot_password') }}">비밀번호 찾기</a></li>
+                        <li><a href="{{ url_for('security.register') }}">회원가입</a></li>
                     </ul>
                 {% else %}
                     <ul class="nav navbar-nav navbar-right">
