+++ setupbox/web_server/templates/security/forgot_password.html
@@ -1,9 +1,12 @@
-{% from "security/_macros.html" import render_field_with_errors, render_field %}
-{% include "security/_messages.html" %}
-<h1>Send password reset instructions</h1>
-<form action="{{ url_for_security('forgot_password') }}" method="POST" name="forgot_password_form">
-  {{ forgot_password_form.hidden_tag() }}
-  {{ render_field_with_errors(forgot_password_form.email) }} 
-  {{ render_field(forgot_password_form.submit) }}
-</form>
-{% include "security/_menu.html" %}
+{% extends 'base.html' %}
+{% block content %}
+    {% from "security/_macros.html" import render_field_with_errors, render_field %}
+    {% include "security/_messages.html" %}
+    <h1>Send password reset instructions</h1>
+    <form action="{{ url_for_security('forgot_password') }}" method="POST" name="forgot_password_form">
+        {{ forgot_password_form.hidden_tag() }}
+        {{ render_field_with_errors(forgot_password_form.email) }}
+        {{ render_field(forgot_password_form.submit) }}
+    </form>
+    {% include "security/_menu.html" %}
+{% endblock %}
